make column metadata functionality opt-in
